Add an option to install dependencies using Poetry Bundle plugin (#28)
This PR adds a new flag to the dependency installation script, --method, which controls the way in which dependencies are installed. By default, the script will still use the old Pip+poetry export behavior. If --method is set to bundle, the Poetry Bundle plugin will be used. Theoretically this should be the preferred option, as it allows Poetry to do all dependency resolution work at install time instead of involving Pip. However, it requires the newly released Poetry 1.2.0 and the Bundle plugin to be installed. This may become the default option in the future if it doesn't cause issues for anyone.

Dependency detection is now done using poetry show instead of by parsing the requirements.txt provided by poetry export. This is a simpler option, and makes it possible to completely remove uses of poetry export in the future.

Fixes #27