Use virtual environments to manage dependencies (#43)
We now create virtual environments in the install/ directory for each package. This fixes issues related to later versions of pip, which would install packages in a "dist-packages" directory, since it thought we were doing a global install. Nodes are also now executed within this virtual environment.

Additionally, the "bundle" dependency installation method is now the only supported method. This was always the better strategy, but I made it opt-in because it relied on a plugin and a (at the time) very recent version of Poetry. These days, I think the reduced complexity and better resiliency makes this new dependency worthwhile.

Note that even though we now use virtual environments, package dependencies are still not isolated from one another. This is because packages need to be able to import each other. We also have to turn on system site packages, so system dependencies are not isolated either. This is because ROS relies on certain system dependencies being available.

Fixes #36